+++ docs/project/changelog.md
@@ -21,6 +21,14 @@ myst:
   to .pyc files
   {pr}`3253`
 
+- Added `pyodide create-zipfile` CLI command that creates a zip file of a directory.
+  This command is hidden by default since it is not intended for use by end users.
+  {pr}`3411`
+
+- {{ Fix }} Fixed a bug where `pyodide build` would fail on package that use CMake,
+  when run multiple times.
+  {pr}`3445`
+
 ### Package Loading
 
 - {{ Fix }} Fix incorrect error message when loading a package